feat: merge vsl staking to vsl core
